requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.3
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Django==4.1.7
django-appconf==1.0.6
django-archive==0.2.0
django-bootstrap-modal-forms==3.0.4
django-compressor==4.4
django-cors-headers==4.3.1
django-cryptography==1.1
django-db-logger==0.1.13
django-environ==0.11.2
django-prettyjson==0.4.1
django-recaptcha==3.0.0
django-tables2==2.7.0
django-tinymce==4.0.0
django-user-agents==0.4.0
djangorestframework==3.15.1
dnspython==2.6.1
et-xmlfile==1.1.0
idna==3.7
isodate==0.6.1
kombu==5.3.7
lxml==5.2.2
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycparser==2.22
pycrypto==2.6.1
pycryptodomex==3.20.0
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-memcached==1.62
pytz==2024.1
rcssmin==1.1.1
redis==5.0.5
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.5.0
standardjson==0.3.1
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
xmltodict==0.13.0
zeep==4.2.1
